fix "sign_integer" function bug and calibrate current sensor offset

+++ User_acmsimcv5/utility.c
@@ -12,8 +12,7 @@ REAL sign(REAL x){
    return (x > 0) - (x < 0);
 }
 
-
-REAL sign_integer(int32 x){
+int32 sign_integer(int32 x){
     return (x > 0) - (x < 0);
 }
